You can now create new processes.

+++ ImageTools.Core/Selection/ProcessStepRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
@@ -7,10 +8,14 @@ namespace ImageTools.Core.Selection
 {
     public class ProcessStepRepository
     {
+        private readonly string _configsFolderPath;
+
         public List<ProcessConfigurationFile> Configs { get; }
 
         public ProcessStepRepository(string configsFolderPath)
         {
+            _configsFolderPath = configsFolderPath;
+
             string[] files = Directory.GetFiles(configsFolderPath);
 
             var configs = new List<ProcessConfigurationFile>();
@@ -26,11 +31,45 @@ namespace ImageTools.Core.Selection
             Configs = configs;
         }
 
+        public string GetSafeFilename(string name)
+        {
+            var firstAttempt = Path.Combine(_configsFolderPath, name + ".json");
+
+            if(!File.Exists(firstAttempt))
+            {
+                return firstAttempt;
+            }
+
+            return Path.Combine(_configsFolderPath, name + "-" + string.Concat(Guid.NewGuid().ToString("N").Take(5)) + ".json");
+        }
+
+        public bool IsSafeId(string name)
+        {
+            return Configs.All(c => c.Id.ToLower() != name.ToLower());
+        }
+
         public void SaveProcessConfigurationFile(ProcessConfigurationFile file)
         {
             var text = JsonConvert.SerializeObject(file, Formatting.Indented);
 
+            if(!File.Exists(file.Filename))
+            {
+                throw new Exception($"File {file.Filename} does not exist.");
+            }
+
             File.WriteAllText(file.Filename, text);
+
+            // Replace config in list with updated one.
+            Configs[Configs.FindIndex(c => c.Id.Equals(file.Id, StringComparison.OrdinalIgnoreCase))] = file;
+        }
+
+        public void CreateProcessConfigurationFile(ProcessConfigurationFile file)
+        {
+            var text = JsonConvert.SerializeObject(file, Formatting.Indented);
+
+            File.WriteAllText(file.Filename, text);
+
+            Configs.Add(file);
         }
     }
